+++ ui/src/app/components/login-gov-ial2-notification.tsx
@@ -3,11 +3,10 @@ import * as React from 'react';
 import colors from 'app/styles/colors';
 import {reactStyles} from 'app/utils';
 import {getAccessModuleStatusByName} from 'app/utils/access-utils';
-import {useNavigation} from 'app/utils/navigation';
 import {profileStore, serverConfigStore, useStore} from 'app/utils/stores';
 import {environment} from 'environments/environment';
 import {AccessModule} from 'generated/fetch';
-import {Button} from './buttons';
+import {Button, LinkButton} from './buttons';
 import {FlexRow} from './flex';
 import {AlarmExclamation} from './icons';
 
@@ -67,19 +66,20 @@ const styles = reactStyles({
 
 
 const LoginGovIAL2Notification = () => {
-  const [navigate, ] = useNavigation();
   return <FlexRow data-test-id='ial2-notification' style={styles.box}>
     <AlarmExclamation style={styles.icon}/>
     <div style={styles.text}>Please verify your identity by 10/27/2021.
     </div>
-    <Button
+    <LinkButton
+        path={
+          environment.enableDataAccessRequirements
+              ? 'data-access-requirements'
+              : '?workbenchAccessTasks=true'
+        }
         type='primary'
-        style={styles.button}
-        onClick={() => environment.enableDataAccessRequirements ?
-            navigate(['data-access-requirements']) :
-            location.replace('/?workbenchAccessTasks=true')}>
+        style={styles.button}>
       <div style={styles.buttonText}>LEARN MORE</div>
-    </Button>
+    </LinkButton>
   </FlexRow>;
 };
